bytebelt/requirements.txt
Django==5.0.6
autopep8
beautifulsoup4
djangorestframework==3.15.1
sqlparse==0.5.0
asgiref==3.8.1
django-filter==24.2
Markdown==3.6
psycopg2==2.9.1
psycopg2-binary==2.9.1
pillow==10.3.0
urllib3
requests
django-dump-die==0.1.7
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-decouple==3.8
django-tailwind==3.8.0
python-slugify==8.0.4
redis==5.0.6
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13